consolidate subresources
